-get solution by tutorial id -save solution from user

+++ src/main/java/vn/hanu/fit/ss2codecamp/repositories/RoleRepository.java
@@ -3,11 +3,11 @@ package vn.hanu.fit.ss2codecamp.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import vn.hanu.fit.ss2codecamp.entities.Role;
-import vn.hanu.fit.ss2codecamp.models.ERole;
+import vn.hanu.fit.ss2codecamp.Enums.ERole;
 
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     Optional<Role> findByName(ERole name);
 }
